Minor changes
Added a method to change authentication manually so that (like in my
case) if a user authenticates to see a screen and then navigates away
from that screen, you can change the authentication factor so the user
is no longer authenticated. I also changed the key in the NSLOG to
actually log the password.